Fix altering db (#17)
Critical bug fix : the full database was dropped if the message table
didn't exist.

This bug is only present in v0.6.0, and only happen when a message is
pushed to the Doctrine repository (i.e. calling Flysystem's `write*` or
`delete*` methods) and the message table doesn't exist (i.e. it's the
first time one of those methods are called in a project).

Technical details: the behavior of Doctrine's `AbstractSchemaManager`
changed from v3 to v4, it now needs the full schema instead of a partial
one containing only the table we want to create.